최신 글 보기 내부 DB 작성 및 채팅방 오류 수정, 홈화면 수정

+++ app/src/main/java/com/with/app/data/AdapterPassData.kt
@@ -6,5 +6,5 @@ data class AdapterPassData(
     var otherName: String? = null,
     var otherProfile: String? = null,
     var chatRoomId: String? = null,
-    var boardIdx: Long = 0
+    var boardIdx: Int = 0
 )
